S4 Filter Implementation
So this is basically an edge-detection algorithm. The way I make it work
is just to do a SafetyCheck when working with the values. This doesn't
change how the filter works, mind, as I did test it using a custom
Sharpen Filter which, too, worked with negative values. Both times
worked like a charm.

We do have a normalization function normalizing everything, but it does
nothing in either casses (be it Cheer or Lena) as we are working on the
same 0-255 scale.

If you try to normalize it with more crazy values (say, -300 to 287),
it'll go crazy and return 0s. Well, the math checks out since it's
0.3981i2739123-something. Basically nothing.