Attempted to Debug Raleigh
Debug success... sort of. I checked every single piece of math in there,
it -does- check out correctly. The only bugger is when they want us to
rise to the power of 0.5, which is equivalent of making a square root
operation.

The gist is, nearly universally you get negative numbers to root, which
is impossible unless you leap into imaginary numbers (something Wolfram
will do for you if you want to check the math with it). Since I don't
KNOW if imaginary numbers should be used in this context or not, I'm
leaving this method as is. Worst to worst we'll fix it before his eyes
on Monday, I'm gone done and fuck-tired of this.

Now for MSE...