MSE and PSNR Implementation
Mean Square Error increases the more deviant the picture is from the
original.

PSNR rises the closer the image is to the original, meaning the less
noise there is in the 'signal'. "Infinity" is a correct value for images
that are nearly the same as the original, meaning we've got 0 noise in
there.

Raleigh can still fuck itself in the ass.