Bug Hunt Report #2 - Raleigh Fix (sortakinda)
Added the (mostly) correct alpha factor calculation - it follows the
formula 1:1, so really there's little chance for deviation. The only
problematic variable is "H(0)", which I couldn't find any explanation
for what it is.

And of course, this doesn't nearly solve the problem of getting
negatives while square-rooting, which throws us NANs. We're one step
closer but we're STILL lacking data and I want to bug him about it
tomorrow. I feel we're -this- freakin' close to solving this bloody
riddle.