Finished implementing filtering.
All forms of filtering now work. The values are still to-be-adjusted
here and there, but that depends on exactly what sort of filter we want
to implement. Usually the best results can be seen when fiddling with
the center of the screen, since this is where majority of our colour
data is. Edges of the screen are near null so do not hold much of an
impact.