Began work on Region Growing algorithm.
So, how does it work.

First of all, I'm discussing a single region. To make it multi-region,
just copy the algorithm, since the basic principle is the same.

We're going to need two tables. A memory table and a region inclusion
table. The first table remembers which pixels we've checked and the
other one is updated on the go with the pixels we've checked and jumped
into, in other words, the region we've grown.
The user inputs a seed pixel (we need to start somewhere!) and a
threshold value (which is a vector between all three channels). We start
at that pixel, adding it to our memory and region tables. Then, we check
the pixel's neighbours. Is the left pixel within the threshold? If so,
jump into it and mark it on both tables. Then, in that pixel, repeat the
operation. Say that the next pixel is not within threshold. Don't jump
into it, but do mark it on the memory map so we don't loop endlessly.
Recursively, we're going to essentially grow out our region like that
until we hit all the pixels that don't fit within the threshold, i.e.
the region barrier.

And that's the Region Growth Segmentation in a nutshell. It's also quite
easy to implement, however the only problem I'm facing right now is how
to figure out that we've stopped - as in, that there's no more search
algorithms running around, that the region is full, complete and cannot
be grown further. If I figure out that, we've got the algorithm.

Simple!